+++ src/polymesh/low_level_api.hh
@@ -188,7 +188,7 @@ struct low_level_attribute_api
     static int offset_attribute_dataptr()
     {
         MeshT mesh;
-        auto attr = mesh.vertices().make_attribute_with_default(0.f);
+        auto attr = mesh.vertices().make_attribute_(0.f);
         auto offset = int(size_t(&(attr.mData)) - size_t(&attr));
         return offset;
     }