+++ src/polymesh/ranges.hh
@@ -217,21 +217,17 @@ struct smart_collection : smart_range<smart_collection<mesh_ptr, tag, iterator>,
     /// Ensures that a given number of primitives can be stored without reallocation
     void reserve(int capacity) const;
 
-    /// Creates a new primitive attribute
+    /// Creates a new primitive attribute (optionally with a default value)
     template <class AttrT>
-    attribute<AttrT> make_attribute() const;
-    /// Creates a new primitive attribute with a given default value
+    attribute<AttrT> make_attribute(AttrT const& def_value = AttrT()) const;
     template <class AttrT>
-    attribute<AttrT> make_attribute_with_default(AttrT const& def_value) const;
+    [[deprecated("use make_attribute instead")]] attribute<AttrT> make_attribute_with_default(AttrT const& def_value) const;
     /// Creates a new primitive attribute and copies the given data
     template <class AttrT>
     attribute<AttrT> make_attribute_from_data(std::vector<AttrT> const& data) const;
     /// Creates a new primitive attribute and copies the given data
     template <class AttrT>
     attribute<AttrT> make_attribute_from_data(AttrT const* data, int cnt) const;
-    /// Creates a new primitive attribute and initializes it with f(h) for each handle h
-    template <class FuncT, class AttrT = tmp::decayed_result_type_of<FuncT, handle>>
-    attribute<AttrT> make_attribute(FuncT&& f, AttrT const& def_value = AttrT()) const;
     /// same as make_attribute(f, def_value)
     template <class FuncT, class AttrT = tmp::decayed_result_type_of<FuncT, handle>>
     attribute<AttrT> map(FuncT&& f, AttrT const& def_value = AttrT()) const;