+++ Changelog.md
@@ -9,3 +9,4 @@
 * `Mesh::copy` and `Mesh::create` now work on `pm::unique_ptr`
 * `properties.hh` was renamed to `view.hh`
 * `algorithms/properties.hh` was moved to `properties.hh`
+* deprecated `make_attribute_with_default` and moved functionality to `make_attribute`